requirements.txt
Babel==1.3
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==2.5.1
SQLAlchemy==0.9.6
WebOb==1.2
chakert==0.1
iktomi==0.4.1
iktomi.cms==0.1
lxml==3.3.5
python-memcached==1.53
six==1.7.3
